requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
pkg-resources==0.0.0
pyFirmata==1.1.0
pyserial==3.5
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
